requirements.txt
autopep8==1.6.0
beautifulsoup4==4.10.0
breadability==0.1.20
cachetools==5.0.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.3
docopt==0.6.2
EbookLib==0.17.1
Flask==2.0.2
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
lxml==4.7.1
MarkupSafe==2.0.1
networkx==2.6.3
nltk==3.6.7
numpy==1.22.2
pdfminer==20191125
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycountry==22.1.10
pycryptodome==3.14.1
pyparsing==3.0.7
python-docx==0.8.11
rake-nltk==1.0.6
regex==2022.1.18
requests==2.27.1
rsa==4.8
scipy==1.8.0
six==1.16.0
soupsieve==2.3.1
sumy==0.9.0
toml==0.10.2
tqdm==4.62.3
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
youtube-transcript-api==0.4.3
